+++ publish/src/commands/deploy/deploy-dapp-utils.js
@@ -2,13 +2,15 @@
 
 const { gray } = require('chalk');
 
-module.exports = async ({ account, addressOf, deployer, readProxyForResolver }) => {
+module.exports = async ({ account, addressOf, deployer }) => {
 	console.log(gray(`\n------ DEPLOY DAPP UTILITIES ------\n`));
+
+	const { ReadProxyAddressResolver } = deployer.deployedContracts;
 
 	await deployer.deployContract({
 		name: 'SynthUtil',
 		deps: ['ReadProxyAddressResolver'],
-		args: [addressOf(readProxyForResolver)],
+		args: [addressOf(ReadProxyAddressResolver)],
 	});
 
 	await deployer.deployContract({
